Time: 0 ms (100.00%) | Memory: 19.8 MB (23.42%) - LeetSync

+++ 745-find-smallest-letter-greater-than-target/find-smallest-letter-greater-than-target.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
     char nextGreatestLetter(vector<char>& letters, char target) {
-        int intCharTarget = (int)target;
+        int intCharTarget = target;
         for (char ch : letters) {
-            int char1 = (int)ch;
+            int char1 = ch;
             if (char1 > intCharTarget) {
                 return ch;
             }
